Introduce Path::Tiny in FileUtils:
To make the transition smoother, I'm starting to first introduce
Path::Tiny into FileUtils to replace the methods in it with
delegation to the appropriate methods in Path::Tiny.

With time, we will simply use Path::Tiny directly.

The first chnge is `open_file`, which will now open a file using
Path::Tiny's `filehandle`.

Dancer2::Handler::File can use Path::Tiny::path()'s
`open_r` directly instead of this more generic methood.

Dancer2::Template::Simple is trickier, because it needs to be
able to load scalarrefs, which Path::Tiny does not support. This
might need to be moved to a utility function, if we use it in
multiple places, but I doubt that.